move interdependent vars into new pico/platform/common.h, and deduplicate some other platform code
